+++ MagusLib/MagusLib/Ertek/TamadoErtekHatarozo.cs
@@ -5,10 +5,10 @@ using MagusLib.KarakterKeszites;
 
 namespace MagusLib.Ertek
 {
-    public class TamadoErtekHatarozo : IHatarozo<uint>
+    public class TamadoErtekHatarozo : IHatarozo<int>
     {
         private readonly KorAlapertekModosito korAlapertekModosito = new KorAlapertekModosito();
-        public uint Hataroz(IKarakter karakter)
+        public int Hataroz(IKarakter karakter)
         {
             try
             {
@@ -16,15 +16,15 @@ namespace MagusLib.Ertek
                 korAlapertekModosito.Karakter = karakter;
 
 
-                uint ertek = (uint)Math.Max((karakter.Ugyesseg + korAlapertekModosito.Ugyesseg - 10), 0);
+                int ertek = Math.Max((karakter.Ugyesseg + korAlapertekModosito.Ugyesseg - 10), 0);
 
-                ertek += (uint)Math.Max((karakter.Gyorsasag + korAlapertekModosito.Gyorsasag - 10), 0);
+                ertek += Math.Max((karakter.Gyorsasag + korAlapertekModosito.Gyorsasag - 10), 0);
 
-                ertek += (uint)Math.Max((karakter.Ero + korAlapertekModosito.Ero - 10), 0);
+                ertek += Math.Max((karakter.Ero + korAlapertekModosito.Ero - 10), 0);
 
                 ertek += Allandok.TE_ALAP[karakter.Alkaszt];
 
-                ertek += (uint)karakter.Szint.TeSzint;
+                ertek += karakter.Szint.TeSzint;
 
                 return ertek;
             }
